requirements.txt
numpy==1.21.5
pandas==1.4.4
umap-learn==0.5.3
scikit-learn==1.0.2
matplotlib==3.5.2
scipy==1.7.3
altair==4.2.0
attrs==22.1.0
blinker==1.5
cachetools==5.2.0
charset-normalizer==2.1.1
click==8.1.3
commonmark==0.9.1
decorator==5.1.1
entrypoints==0.4
fonttools==4.25.0
gitdb==4.0.9
GitPython==3.1.27
idna==3.4
importlib-metadata==4.12.0
Jinja2==3.1.2
jsonschema==4.16.0
llvmlite==0.38.0
munkres==1.1.4
Pillow==9.2.0
ply==3.11
protobuf==3.20.1
pyarrow==9.0.0
pydeck==0.8.0b3
Pygments==2.13.0
Pympler==1.0.1
PyQt5==5.15.7
pyrsistent==0.18.1
requests==2.28.1
rich==12.5.1
semver==2.13.0
smmap==5.0.0
streamlit==1.13.0
toolz==0.12.0
typing_extensions==4.3.0
tzdata==2022.4
tzlocal==4.2
urllib3==1.26.12
validators==0.20.0
watchdog==2.1.9
wincertstore==0.2
zipp==3.8.1
xgboost==1.5.0